RF_03.003 | BreadCrumbsDropDownMenuTest> deleting class BreadCrumbsDropDownMenuTest aws deleted and moved to FreestyleProjectTest
